requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

transformers>=4.42.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
sentence-transformers==2.3.1

faiss-cpu==1.7.4
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4

datasets==2.14.6
tokenizers>=0.15.0
huggingface-hub>=0.34.0
safetensors>=0.4.1
accelerate==0.24.1
bitsandbytes==0.41.3

requests==2.31.0
pyyaml==6.0.1
fsspec>=2023.1.0,<=2023.10.0

regex==2023.10.3
tqdm==4.66.1
xxhash==3.4.1
multiprocess==0.70.15
dill==0.3.7
pyarrow==14.0.1

langchain>=0.1.17
langchain-community>=0.0.29
